clear institution fields no longer being updated by the hmda-platform service
